pyproject.toml
[project]
name = "todo-list"
version = "0.1.0"
description = ""
authors = [
    {name = "Khushi Bansal",email = "bansalkhushi880@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "bcrypt (==4.0.1)",
]

[tool.poetry]
packages = [{include = "todo_list", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
